Refactor type mapper
